notification sent for each message, to mentor and mentee and partner
With regards to email notification, I would like a notification sent for each message, to mentor and mentee and partner. Each can select email or text and both should receive the notification right away. We can include a small line at the end saying click on your profile to change notifications to email (when receiving a text) or change to text (when receiving an email) to remind them of both options.  Emails sent through Sendgrid and SMS through Twilio.  Both interfaces should be already created.