https://github.com/hack4impact-uiuc/mentee/issues/551
Admin user is able to update the profile picture of the Mentor, and the Mentees, mostly when they have no picture.  In some cases, admin users faced an issue, and the image was not being saved correctly in the database, in most cases it worked fine.  Need cost for finding the root cause and fixing it. https://github.com/hack4impact-uiuc/mentee/issues/551